Separation of business logic into own component (DLL) form UI implementation. Helps adaptability in other applications.
