+++ SOAPMessageBuilder.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SOAPMessageBuilder", "SOAPM
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XMLGenerator", "XMLGenerator\XMLGenerator.csproj", "{ACA48E89-D110-4C24-B90F-8775543E5546}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BusinessLogic", "BusinessLogic\BusinessLogic.csproj", "{4CA95D81-F8B2-4248-AAE2-4C2C62E6E621}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{ACA48E89-D110-4C24-B90F-8775543E5546}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ACA48E89-D110-4C24-B90F-8775543E5546}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ACA48E89-D110-4C24-B90F-8775543E5546}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4CA95D81-F8B2-4248-AAE2-4C2C62E6E621}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4CA95D81-F8B2-4248-AAE2-4C2C62E6E621}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4CA95D81-F8B2-4248-AAE2-4C2C62E6E621}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4CA95D81-F8B2-4248-AAE2-4C2C62E6E621}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
